requirements.txt
pygame==2.5.2
scapy==2.5.0
cryptography==42.0.0
pycryptodome==3.19.0
colorama==0.4.6
prompt-toolkit==3.0.43
python-dotenv==1.0.0
typing-extensions==4.9.0
pytest==7.4.3
pytest-cov==4.1.0
opencv-python==4.8.1.78
numpy==1.26.2
Pillow==12.1.0